requirements.txt
# Core AI and Vector Search
langchain
langchain-google-genai
langchain-community
sentence-transformers
pinecone-client # Replaces faiss-cpu

# Document Loaders & HTTP requests
pypdf
requests

# API Framework
fastapi
uvicorn[standard]

# For environment variables
python-dotenv